+++ src/errors.ts
