+++ src/index.ts
@@ -1,5 +1,5 @@
 // NOTE: do NOT export `Form` here. Next.js can't properly tree shake the client
 // bundle if there are client & server imports in a single import statement.
-export * from './error';
+export * from './errors';
 export * from './handle';
 export * from './responses';
